requirements.txt
# Core OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Numerical & data
numpy>=1.24.0
pandas>=2.0.0

# Metrics
jiwer>=3.0.0
editdistance>=0.6.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML utilities
scikit-learn>=1.3.0

# Augmentation (optional)
imgaug>=0.4.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
